+++ app/Driver.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App;
+
+class Driver
+{
+    public ?Licence $licence = null;
+}
