+++ app/Invoices/Credit/Bill.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Public\Invoices\Credit;
+
+class Bill
+{
+
+}
