+++ app/Invoices/Subscription/Bill.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Public\Invoices\Subscription;
+
+class Bill
+{
+    public function __construct()
+    {
+        $customerProfile = new CustomerProfile();
+
+        $email = new \App\Notifications\Email();
+
+        var_dump($email);
+    }
+}
