+++ app/Invoices/Subscription/CustomerProfile.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Public\Invoices\Subscription;
+
+class CustomerProfile
+{
+
+}
