+++ app/Licence.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App;
+class Licence
+{
+    public ?string $expirationDate = null;
+}
