+++ composer.json
@@ -15,7 +15,7 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": "app/"
         }
     }
 }
