--- public/Driver.php
@@ -1,6 +1,0 @@
-<?php
-
-class Driver
-{
-    public ?Licence $licence = null;
-}
