--- public/Invoices/Credit/Bill.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Public\Invoices\Credit;
-
-class Bill
-{
-
-}
