--- public/Invoices/Invoice/Bill.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Public\Invoices\Invoice;
-
-class Bill
-{
-
-}
