--- public/Invoices/Subscription/CustomerProfile.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Public\Invoices\Subscription;
-
-class CustomerProfile
-{
-
-}
