--- public/Licence.php
@@ -1,6 +1,0 @@
-<?php
-
-class Licence
-{
-    public ?string $expirationDate = null;
-}
