--- public/Notifications/Email.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Public\Notifications;
-
-class Email
-{
-
-}
